requirements.txt
Django>=5.0
gunicorn>=21.0
whitenoise>=6.6
yt-dlp>=2024.12.06
groq==1.1.1
python-dotenv==1.2.2
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
